requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
Pillow==10.0.0
requests==2.31.0
scikit-learn==1.3.1
pandas==2.1.1
numpy==1.25.2
gunicorn==21.2.0
psycopg2-binary==2.9.9